requirements.txt
langchain
python-dotenv
langchain-community
streamlit
openai
langchain_openai